+++ flit/{{cookiecutter.github_repo}}/CHANGELOG.md
@@ -1,5 +1,3 @@
 # Changelog
 
-## Version {{cookiecutter.pkg_version}} (alpha)
-
-- Initial commit
+- [Version {{cookiecutter.pkg_version}}](https://github.com/{{cookiecutter.github_user}}/{{cookiecutter.github_repo}}/releases/tag/v{{cookiecutter.pkg_version}})